% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_stats.R
\name{diagnose_permutation_resolution}
\alias{diagnose_permutation_resolution}
\title{Diagnose Permutation Null Resolution}
\usage{
diagnose_permutation_resolution(null_pair)
}
\arguments{
\item{null_pair}{Output from \code{permute_celltype_communication()}.}
}
\value{
A \code{LogicCommPermutationDiagnostic} list.
}
\description{
Diagnose Permutation Null Resolution
}
