% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareLogicGroups.R
\name{.append_lr_metadata}
\alias{.append_lr_metadata}
\title{Append LR metadata while preserving result row order}
\usage{
.append_lr_metadata(result, lr_db = NULL)
}
\description{
Append LR metadata while preserving result row order
}
\keyword{internal}
