% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REO_matrix.R
\name{.calc_REO_bpcells}
\alias{.calc_REO_bpcells}
\title{BPCells streaming variant}
\usage{
.calc_REO_bpcells(
  bp_mat,
  lr_genes,
  rank_threshold,
  chunk_size,
  return_rank = FALSE,
  rank_output = "percentile",
  anchor_genes = NULL,
  verbose = TRUE
)
}
\description{
BPCells streaming variant
}
\keyword{internal}
