% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_participation.R
\name{.composition_table}
\alias{.composition_table}
\title{Composition of active communication by pathway or L-R pair and direction}
\usage{
.composition_table(lr_table, key = "pathway", extra = character(0))
}
\description{
Composition of active communication by pathway or L-R pair and direction
}
\keyword{internal}
