% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rank_shift.R
\name{.compute_expr_means}
\alias{.compute_expr_means}
\title{Compute mean expression per gene per group}
\usage{
.compute_expr_means(
  sample_list,
  genes,
  group_info,
  case_label,
  ctrl_label,
  layer
)
}
\description{
Compute mean expression per gene per group
}
\keyword{internal}
