% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_communication.R
\name{.differential_subgroup_summary}
\alias{.differential_subgroup_summary}
\title{Per sender -> receiver subgroup differential summary}
\usage{
.differential_subgroup_summary(lr, min_asymmetry, case_label, ctrl_label)
}
\description{
Per sender -> receiver subgroup differential summary
}
\keyword{internal}
