% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulators.R
\name{.has_lr_genes}
\alias{.has_lr_genes}
\title{Check whether an LR row has any genes in a modulator list column}
\usage{
.has_lr_genes(lr_db, col, i)
}
\description{
Check whether an LR row has any genes in a modulator list column
}
\keyword{internal}
