% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_viz.R
\docType{data}
\name{.logiccomm_palettes}
\alias{.logiccomm_palettes}
\title{LogicComm Color Palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
.logiccomm_palettes
}
\description{
LogicComm Color Palettes
}
\keyword{internal}
