% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rank_shift.R
\name{.median_norm_rank_per_gene}
\alias{.median_norm_rank_per_gene}
\title{Compute median normalized rank per gene across cells in one sample}
\usage{
.median_norm_rank_per_gene(mat, min_detection_frac)
}
\description{
Compute median normalized rank per gene across cells in one sample
}
\keyword{internal}
