% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_gate.R
\name{.positive_modulator_gate}
\alias{.positive_modulator_gate}
\title{Compute active logic for optional positive modulators}
\usage{
.positive_modulator_gate(genes, reo_mat, mode = c("all", "any", "ignore"))
}
\description{
Compute active logic for optional positive modulators
}
\keyword{internal}
