% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_gate.R
\name{.positive_modulator_gate_v2}
\alias{.positive_modulator_gate_v2}
\title{Compute active logic for optional positive modulators (v2)}
\usage{
.positive_modulator_gate_v2(genes, get_logic, mode = c("all", "any", "ignore"))
}
\description{
Compute active logic for optional positive modulators (v2)
}
\keyword{internal}
