% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REO_matrix.R
\name{.rank_percentile_subset}
\alias{.rank_percentile_subset}
\title{Rank retained genes against a within-cell universe}
\usage{
.rank_percentile_subset(universe_chunk, output_genes)
}
\description{
Rank retained genes against a within-cell universe
}
\keyword{internal}
