% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REO_matrix.R
\name{.select_anchor_genes}
\alias{.select_anchor_genes}
\title{Select anchor genes for REO calculation}
\usage{
.select_anchor_genes(all_genes, anchor_genes = NULL, label = "matrix")
}
\description{
Select anchor genes for REO calculation
}
\keyword{internal}
