% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REO_matrix.R
\name{.select_output_genes}
\alias{.select_output_genes}
\title{Select output genes for REO calculation}
\usage{
.select_output_genes(all_genes, lr_genes, label = "matrix", verbose = TRUE)
}
\description{
Select output genes for REO calculation
}
\keyword{internal}
