% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyLogicConsensus.R
\name{.sparse_to_edges}
\alias{.sparse_to_edges}
\title{Convert a KNN matrix to directed edge index vectors (1-based)}
\usage{
.sparse_to_edges(
  knn_mat,
  remove_self_edges = TRUE,
  edge_weight_mode = c("binary", "weighted")
)
}
\description{
Convert a KNN matrix to directed edge index vectors (1-based)
}
\keyword{internal}
