% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyLogicConsensus.R
\name{.symmetrize_sparse_graph}
\alias{.symmetrize_sparse_graph}
\title{Symmetrize a sparse graph while preserving optional edge weights}
\usage{
.symmetrize_sparse_graph(knn_mat, mode = c("none", "or", "max"))
}
\description{
Symmetrize a sparse graph while preserving optional edge weights
}
\keyword{internal}
