% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_gate.R
\name{.validate_rank_mat}
\alias{.validate_rank_mat}
\title{Validate and align rank matrix to REO matrix}
\usage{
.validate_rank_mat(rank_mat, reo_mat)
}
\description{
Validate and align rank matrix to REO matrix
}
\keyword{internal}
