% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multisample.R
\name{get_comparison}
\alias{get_comparison}
\title{Extract the comparison result from a LogicCommMulti object}
\usage{
get_comparison(x)
}
\arguments{
\item{x}{LogicCommMulti object}
}
\value{
LogicCommResult data frame
}
\description{
Extract the comparison result from a LogicCommMulti object
}
