% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_roles.R
\name{plot_celltype_roles}
\alias{plot_celltype_roles}
\title{Plot Sender-Receiver Role Positioning}
\usage{
plot_celltype_roles(ct_comm, label = TRUE, title = NULL)
}
\arguments{
\item{ct_comm}{Output from \code{summarize_celltype_communication()}.}

\item{label}{Logical; label cell types. Default: \code{TRUE}.}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object.
}
\description{
Plot Sender-Receiver Role Positioning
}
