% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_extensions.R
\name{plot_communication_qc}
\alias{plot_communication_qc}
\title{Plot Communication Quality-Control Landscape}
\usage{
plot_communication_qc(ct_comm, title = NULL)
}
\arguments{
\item{ct_comm}{Output from summarize_celltype_communication().}

\item{title}{Optional title.}
}
\value{
ggplot2 object.
}
\description{
Plot Communication Quality-Control Landscape
}
