% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_plots.R
\name{plot_communication_range_summary}
\alias{plot_communication_range_summary}
\title{Plot Communication Range Summary}
\usage{
plot_communication_range_summary(ct_comm)
}
\arguments{
\item{ct_comm}{Output from \code{summarize_celltype_communication()}.}
}
\value{
A ggplot2 object.
}
\description{
Plot Communication Range Summary
}
