% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_extensions.R
\name{print.LogicCommGLMResult}
\alias{print.LogicCommGLMResult}
\title{Print method for LogicCommGLMResult}
\usage{
\method{print}{LogicCommGLMResult}(x, ...)
}
\arguments{
\item{x}{LogicCommGLMResult object.}

\item{...}{Ignored.}
}
\description{
Print method for LogicCommGLMResult
}
