% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_REO_matrix.R
\name{print.LogicCommREOResult}
\alias{print.LogicCommREOResult}
\title{Print method for LogicCommREOResult}
\usage{
\method{print}{LogicCommREOResult}(x, ...)
}
\arguments{
\item{x}{LogicCommREOResult object.}

\item{...}{Ignored.}
}
\description{
Print method for LogicCommREOResult
}
