% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rank_shift.R
\name{print.RankShiftResult}
\alias{print.RankShiftResult}
\title{Print method for RankShiftResult}
\usage{
\method{print}{RankShiftResult}(x, n = 10, ...)
}
\arguments{
\item{x}{RankShiftResult}

\item{n}{Rows to display. Default: 10.}

\item{...}{ignored}
}
\description{
Print method for RankShiftResult
}
