% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoLabelLogicClusters.R
\name{AutoLabelLogicClusters}
\alias{AutoLabelLogicClusters}
\title{Auto-Annotate Cell Clusters Driving Logic Consensus Hotspots}
\usage{
AutoLabelLogicClusters(
  reo_mat,
  cell_labels,
  lr_db = lr_pairs_human,
  lr_pairs = NULL,
  min_sender_frac = 0.1,
  min_receiver_frac = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{reo_mat}{REO binary matrix.}

\item{cell_labels}{Named character vector or Seurat object.}

\item{lr_db}{LR database.}

\item{lr_pairs}{Character vector of L-R pairs to analyze.}

\item{min_sender_frac}{Minimum fraction for sender.}

\item{min_receiver_frac}{Minimum fraction for receiver.}

\item{verbose}{Logical.}
}
\value{
A data frame with cluster enrichment stats.
}
\description{
Auto-Annotate Cell Clusters Driving Logic Consensus Hotspots
}
