% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_plots.R
\name{plot_celltype_role_dotplot}
\alias{plot_celltype_role_dotplot}
\title{Plot Cell-Type Role Dotplot}
\usage{
plot_celltype_role_dotplot(ct_comm, scaled = TRUE, title = NULL)
}
\arguments{
\item{ct_comm}{Cell-type communication object.}

\item{scaled}{If TRUE, z-score role scores before plotting.}

\item{title}{Optional title.}
}
\value{
ggplot2 object.
}
\description{
Plot Cell-Type Role Dotplot
}
