% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_plots.R
\name{plot_lr_bubble_advanced}
\alias{plot_lr_bubble_advanced}
\title{Advanced Bubble Plot of L-R Logic Events}
\usage{
plot_lr_bubble_advanced(
  ct_comm,
  senders = NULL,
  receivers = NULL,
  top_n_per_pair = 5,
  top_n_pathways = 12
)
}
\arguments{
\item{ct_comm}{Output from \code{summarize_celltype_communication()}.}

\item{senders}{Optional sender filter.}

\item{receivers}{Optional receiver filter.}

\item{top_n_per_pair}{Number of rows per sender-receiver pair.}

\item{top_n_pathways}{Number of pathways (facets) to show, ranked by total LCS.}
}
\value{
A ggplot2 object.
}
\description{
Advanced Bubble Plot of L-R Logic Events
}
