% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_plots.R
\name{plot_lr_bubble_by_celltype}
\alias{plot_lr_bubble_by_celltype}
\title{Plot L-R Bubbles by Cell Type}
\usage{
plot_lr_bubble_by_celltype(
  ct_comm,
  sender = NULL,
  receiver = NULL,
  top_n = 30,
  active_only = TRUE,
  color_by = c("pathway", "lcs", "n_active_edges", "ligand_active_frac_sender",
    "receptor_active_frac_receiver"),
  title = NULL
)
}
\arguments{
\item{ct_comm}{Output from \code{summarize_celltype_communication()}.}

\item{sender}{Optional sender filter.}

\item{receiver}{Optional receiver filter.}

\item{top_n}{Number of rows to show.}

\item{active_only}{Whether to show active candidate rows only.}

\item{color_by}{Field used for color.}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot L-R Bubbles by Cell Type
}
