% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_plots.R
\name{plot_lr_evidence}
\alias{plot_lr_evidence}
\title{Plot Evidence for One Cell-Type L-R Event}
\usage{
plot_lr_evidence(ct_comm, sender, receiver, lr_pair, title = NULL)
}
\arguments{
\item{ct_comm}{Output from \code{summarize_celltype_communication()}.}

\item{sender}{Sender cell type.}

\item{receiver}{Receiver cell type.}

\item{lr_pair}{L-R pair.}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot Evidence for One Cell-Type L-R Event
}
