% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_plots.R
\name{plot_pathway_dominance}
\alias{plot_pathway_dominance}
\title{Plot Pathway Dominance}
\usage{
plot_pathway_dominance(ct_comm, metric = "sum_lcs", top_n = 15, title = NULL)
}
\arguments{
\item{ct_comm}{Cell-type communication object.}

\item{metric}{Pathway metric, usually \code{"sum_lcs"}.}

\item{top_n}{Number of pathways to display.}

\item{title}{Optional title.}
}
\value{
ggplot2 object.
}
\description{
Plot Pathway Dominance
}
