% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_plots.R
\name{plot_pathway_network}
\alias{plot_pathway_network}
\title{Plot a Pathway-Filtered Cell-Type Network}
\usage{
plot_pathway_network(
  ct_comm,
  pathway,
  metric = "sum_lcs",
  top_n_edges = 20,
  title = NULL
)
}
\arguments{
\item{ct_comm}{Cell-type communication object.}

\item{pathway}{Pathway name to show.}

\item{metric}{Pathway metric.}

\item{top_n_edges}{Number of edges to show.}

\item{title}{Optional title.}
}
\value{
ggplot2 object.
}
\description{
Plot a Pathway-Filtered Cell-Type Network
}
