% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyLogicConsensus.R
\name{print.LCSVector}
\alias{print.LCSVector}
\title{Print method for LCSVector}
\usage{
\method{print}{LCSVector}(x, ...)
}
\arguments{
\item{x}{LCSVector}

\item{...}{ignored}
}
\description{
Print method for LCSVector
}
