% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_extensions.R
\name{print.LogicCommDynamics}
\alias{print.LogicCommDynamics}
\title{Print method for LogicCommDynamics}
\usage{
\method{print}{LogicCommDynamics}(x, ...)
}
\arguments{
\item{x}{LogicCommDynamics object.}

\item{...}{Ignored.}
}
\description{
Print method for LogicCommDynamics
}
