% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_extensions.R
\name{print.LogicCommFindings}
\alias{print.LogicCommFindings}
\title{Print method for LogicCommFindings}
\usage{
\method{print}{LogicCommFindings}(x, ...)
}
\arguments{
\item{x}{LogicCommFindings object.}

\item{...}{Ignored.}
}
\description{
Print method for LogicCommFindings
}
