% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multisample.R
\name{print.LogicCommMulti}
\alias{print.LogicCommMulti}
\title{Print method for LogicCommMulti}
\usage{
\method{print}{LogicCommMulti}(x, ...)
}
\arguments{
\item{x}{LogicCommMulti object}

\item{...}{ignored}
}
\description{
Print method for LogicCommMulti
}
