% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareLogicGroups.R
\name{print.LogicCommResult}
\alias{print.LogicCommResult}
\title{Print method for LogicCommResult}
\usage{
\method{print}{LogicCommResult}(x, n = 10, ...)
}
\arguments{
\item{x}{LogicCommResult data frame.}

\item{n}{Number of rows to print. Default: 10.}

\item{...}{Ignored.}
}
\description{
Print method for LogicCommResult
}
