% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_gate.R
\name{print.LogicGateResult}
\alias{print.LogicGateResult}
\title{Print method for LogicGateResult}
\usage{
\method{print}{LogicGateResult}(x, ...)
}
\arguments{
\item{x}{LogicGateResult object.}

\item{...}{ignored.}
}
\description{
Print method for LogicGateResult
}
