% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_viz.R
\name{theme_logiccomm}
\alias{theme_logiccomm}
\title{Premium LogicComm Visual Theme}
\usage{
theme_logiccomm(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Premium LogicComm Visual Theme
}
